pyproject.toml
[project]
name = "neko-agent"
version = "1.0.0"
description = "Provider-agnostic Python agent framework inspired by Claude Code architecture"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",     # OpenAI-compatible APIs (OpenAI, MiniMax, Kimi, DeepSeek, Qwen)
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",         # Model Context Protocol
    "rich>=13.0",         # Terminal UI
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.30.0"]    # For Claude models
gemini = ["google-genai>=1.0.0"]     # For Gemini models
all = ["anthropic>=0.30.0", "google-genai>=1.0.0"]  # All providers
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
agent = "neko_agent.__main__:main"
